refactor: residual helpers + useApiKeySave, client→shell (#3501 Phase 1s)
